refactor: move auth recovery under auth domain
