+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/recovery/dtos/PasswordResetTokenResponseDto.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.adapters.inbound.web.accountRecovery.dtos
+package app.xquare.xquareinfra.adapters.inbound.web.auth.recovery.dtos
 
 import app.xquare.xquareinfra.infrastructure.web.dto.SuccessResponseDto
