+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/recovery/dtos/SendPasswordResetOtpRequestDto.kt
@@ -1,16 +1,15 @@
-package app.xquare.xquareinfra.adapters.inbound.web.accountRecovery.dtos
+package app.xquare.xquareinfra.adapters.inbound.web.auth.recovery.dtos
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.Max
 import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
 
-data class VerifyPasswordResetOtpRequestDto(
+data class SendPasswordResetOtpRequestDto(
     @field:Size(min = 4, max = 15)
     @field:Pattern(
-        regexp = "^[A-Za-z0-9_-]+$",
+        regexp = "^[A-z0-9_-]+$",
     )
     val username: String,
     @field:Min(1000)
@@ -22,7 +21,4 @@ data class VerifyPasswordResetOtpRequestDto(
     val name: String,
     @field:Email
     val email: String,
-    @field:Size(min = 6, max = 6)
-    @field:NotBlank
-    val otp: String,
 )
