+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/recovery/dtos/SendUsernameFindOtpRequestDto.kt
@@ -1,13 +1,11 @@
-package app.xquare.xquareinfra.adapters.inbound.web.accountRecovery.dtos
+package app.xquare.xquareinfra.adapters.inbound.web.auth.recovery.dtos
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.Max
 import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Pattern
-import jakarta.validation.constraints.Size
 
-data class VerifyUsernameFindOtpRequestDto(
+data class SendUsernameFindOtpRequestDto(
     @field:Min(1000)
     @field:Max(3999)
     val studentNumber: Int,
@@ -17,8 +15,4 @@ data class VerifyUsernameFindOtpRequestDto(
     val name: String,
     @field:Email
     val email: String,
-    @field:Pattern(regexp = "^\\d{6}$")
-    @field:Size(min = 6, max = 6)
-    @field:NotBlank
-    val otp: String,
 )
