+++ src/main/kotlin/app/xquare/xquareinfra/adapters/inbound/web/auth/recovery/dtos/VerifyUsernameFindOtpRequestDto.kt
@@ -1,11 +1,13 @@
-package app.xquare.xquareinfra.adapters.inbound.web.accountRecovery.dtos
+package app.xquare.xquareinfra.adapters.inbound.web.auth.recovery.dtos
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.Max
 import jakarta.validation.constraints.Min
+import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Pattern
+import jakarta.validation.constraints.Size
 
-data class SendUsernameFindOtpRequestDto(
+data class VerifyUsernameFindOtpRequestDto(
     @field:Min(1000)
     @field:Max(3999)
     val studentNumber: Int,
@@ -15,4 +17,8 @@ data class SendUsernameFindOtpRequestDto(
     val name: String,
     @field:Email
     val email: String,
+    @field:Pattern(regexp = "^\\d{6}$")
+    @field:Size(min = 6, max = 6)
+    @field:NotBlank
+    val otp: String,
 )
