+++ src/main/kotlin/app/xquare/xquareinfra/application/auth/ports/inbound/recovery/SendPasswordResetOtpUseCase.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.application.accountRecovery.ports.inbound
+package app.xquare.xquareinfra.application.auth.ports.inbound.recovery
 
 interface SendPasswordResetOtpUseCase {
     fun sendPasswordResetOtp(command: SendPasswordResetOtpCommand)
