+++ src/main/kotlin/app/xquare/xquareinfra/application/auth/ports/inbound/recovery/SendUsernameFindOtpUseCase.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.application.accountRecovery.ports.inbound
+package app.xquare.xquareinfra.application.auth.ports.inbound.recovery
 
 interface SendUsernameFindOtpUseCase {
     fun sendUsernameFindOtp(command: SendUsernameFindOtpCommand)
