+++ src/main/kotlin/app/xquare/xquareinfra/application/auth/ports/inbound/recovery/VerifyUsernameFindOtpUseCase.kt
@@ -1,4 +1,4 @@
-package app.xquare.xquareinfra.application.accountRecovery.ports.inbound
+package app.xquare.xquareinfra.application.auth.ports.inbound.recovery
 
 interface VerifyUsernameFindOtpUseCase {
     fun verifyUsernameFindOtp(command: VerifyUsernameFindOtpCommand): VerifyUsernameFindOtpResult
