+++ src/main/kotlin/app/xquare/xquareinfra/application/auth/ports/outbound/UserPersistenceForAuthPort.kt
@@ -9,5 +9,20 @@ interface UserPersistenceForAuthPort {
 
     fun save(user: User): User
 
+    fun findByEmail(email: String): User?
+
+    fun findByStudentNumberAndNameAndEmail(
+        studentNumber: Int,
+        name: String,
+        email: String,
+    ): List<User>
+
+    fun findByUsernameAndStudentNumberAndNameAndEmail(
+        username: String,
+        studentNumber: Int,
+        name: String,
+        email: String,
+    ): User?
+
     fun findByUsername(username: String): User?
 }
